pyproject.toml
[tool.poetry]
name = "graph-quantification"
version = "0.1.0"
description = ""
authors = []
readme = "README.md"
classifiers = ["Topic :: Scientific/Engineering :: Artificial Intelligence"]
packages = [
    { include = "gpn", from = "." },
    { include = "pyblaze", from = "pyblaze" },
    { include = "mdmm_jax", from = "mdmm-jax" },
]

[tool.poetry.dependencies]
python = ">=3.10,<3.13"
torch = { version = "2.5.1", source = "torch" }
torch-geometric = "2.6.1"
pyg_lib = { version = "0.4.0", source = "pyg" }
torch_scatter = { version = "2.1.2", source = "pyg" }
torch_sparse = { version = "0.6.18", source = "pyg" }
networkx = "^3.4.2"
scipy = "^1.14.1"
matplotlib = "^3.9.2"
funcy = "^2.0"
ogb = "^1.3.6"
sacred = "^0.8.5"
seaborn = "^0.13.0"
plotly = "^5.24.1"
pandas = "2.2.2"
tinydb = "^4.8.2"
tinydb-serialization = "^2.2.0"
scikit-learn = "^1.5.2"
numpy = "^1.26.4"
tensorboard = "2.14.1"
pyro-ppl = "^1.9.1"
tensorflow = "2.14.1"
gpflow = "2.9.0"
pyarrow = "^15.0.2"
tensorflow-probability = "0.22.1"
quapy = "^0.1.9"
tensorflow-macos = [{ version = "2.14.1", platform = "darwin" }]
cugraph-cu12 = { version = "24.12.0", source = "pypi-nvidia" }
cudf-cu12 = { version = "24.12.0", source = "pypi-nvidia" }
pylibcudf-cu12 = { version = "24.12.0", source = "pypi-nvidia" }
libcudf-cu12 = { version = "24.12.0", source = "pypi-nvidia" }
pylibcugraphops-cu12 = { version = "24.12.0", source = "pypi-nvidia" }
nx-cugraph-cu12 = { version = "24.12.0", source = "pypi-nvidia" }
igraph = "^0.11.8"
jax = { version = "0.4.34", extras = ["cuda12_local"] }
optax = "^0.2.4"
iteround = "^1.0.4"
python-ternary = "^1.0.8"

[[tool.poetry.source]]
name = "torch"
url = "https://download.pytorch.org/whl/cu124"
priority = "explicit"

[[tool.poetry.source]]
name = "pyg"
url = "https://data.pyg.org/whl/torch-2.5.0+cu124.html"
priority = "explicit"

[[tool.poetry.source]]
name = "pypi-nvidia"
url = "https://pypi.nvidia.com"
priority = "explicit"

[tool.poetry.group.dev.dependencies]
ipykernel = "^6.29.5"
notebook = "^7.3.2"
ipywidgets = "^8.1.5"
pyright = "^1.1.391"
types-tensorflow = "^2.18.0.20241112"
per4m = "^0.1.0"
viztracer = "^1.0.0"

[tool.pyright]
include = ["gpn", "pyblaze"]
pythonVersion = "3.10"
pythonPlatform = "Linux"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
